backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "google-genai>=0.7.2",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
]
